userid

+++ app/models/auth/model/oauth_user.rb
@@ -23,7 +23,7 @@ module Auth
       belongs_to :user, optional: true
       belongs_to :account, -> { where(confirmed: true) }, foreign_key: :identity, primary_key: :identity, inverse_of: :oauth_users, optional: true
 
-      has_many :authorized_tokens, ->(o) { where(appid: o.appid, identity: o.identity) }, primary_key: :uid, foreign_key: :uid, dependent: :delete_all
+      has_many :authorized_tokens, ->(o) { where(appid: o.appid, user_id: o.user_id, identity: o.identity) }, primary_key: :uid, foreign_key: :uid, dependent: :delete_all
       belongs_to :same_oauth_user, ->(o) { where.not(id: o.id) }, class_name: self.name, foreign_key: :unionid, primary_key: :unionid, optional: true
       has_many :same_oauth_users, class_name: self.name, primary_key: :unionid, foreign_key: :unionid
